Revert "Better deployment"
