--- js/packages/botProtocol/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "botprotocol",
-  "version": "1.0.0",
-  "description": "tokens for surface <-> bot communication",
-  "main": "botProtocol.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Nugget Industries",
-  "license": "ISC",
-  "dependencies": {
-    "uuid": "^3.1.0"
-  }
-}
