--- js/packages/botSocket/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "botsocket",
-  "version": "1.0.0",
-  "description": "API for talking to the bot",
-  "main": "botSocket.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Nugget Industries",
-  "license": "ISC",
-  "dependencies": {
-    "botprotocol": "file:../botProtocol"
-  }
-}
