--- js/packages/controller/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "controller",
-  "version": "1.0.0",
-  "description": "API for talking to our controller",
-  "main": "controller.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Nugget Industries",
-  "license": "ISC",
-  "dependencies": {
-    "serialport": "^6.0.4"
-  }
-}
