--- js/packages/deploy/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "deploy",
-  "version": "1.0.0",
-  "description": "For deploying fun things to the robot",
-  "main": "deploy.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Nugget Industries",
-  "license": "ISC",
-  "dependencies": {
-    "remote-exec": "0.0.3",
-    "scp2": "^0.5.0",
-    "yargs": "^11.0.0"
-  }
-}
