--- js/remote/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "remote",
-  "version": "1.0.0",
-  "description": "for the bot",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Nugget Industries",
-  "license": "ISC",
-  "dependencies": {
-    "botprotocol": "file:../packages/botProtocol",
-    "yargs": "^11.0.0"
-  }
-}
