--- js/surface/dashboard/templates/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>NUGboard</title>
-    <link rel="shortcut icon" href="/static/images/nugget.ico" type="image/x-icon">
-    <link rel="stylesheet" type="text/css" href="/static/lib/indicators/css/flightindicators.css">
-    <script src="/static/lib/jquery.js"></script>
-    <script src="/static/lib/indicators/js/jquery.flightindicators.min.js"></script>
-    <script src="/static/lib/socket.io.js"></script>
-    <script src="/static/dashSocket.js"></script>
-  </head>
-  <body>
-    <p>greetings</p><span id="attitude"></span><span id="heading"></span>
-    <button id="connect">connect</button>
-    <button id="disconnect">disconnect</button>
-  </body>
-</html>
