--- js/test/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "test",
-  "version": "1.0.0",
-  "description": "integration testing 4 our bot",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Nugget Industries",
-  "license": "ISC",
-  "dependencies": {
-    "botsocket": "file:../packages/botSocket",
-    "deploy": "file:../packages/deploy",
-    "mocha": "^4.0.1"
-  }
-}
