+++ production/packages/deploy/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "deploy",
+  "version": "1.0.0",
+  "description": "For deploying fun things to the robot",
+  "main": "deploy.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Bobby Martin",
+  "license": "ISC",
+  "dependencies": {
+    "clp": "^4.0.8"
+  }
+}
