+++ production/remote/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "remote",
+  "version": "1.0.0",
+  "description": "for the bot",
+  "main": "botServer.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Bobby Martin",
+  "license": "ISC",
+  "dependencies": {
+    "botprotocol": "file:../packages/botProtocol",
+    "clp": "^4.0.8"
+  }
+}
