+++ production/surface/dashboard/static/dashSocket.js
